Move reinforcement spend endpoint under /skills/
