+++ front/pages/api/w/[wId]/skills/reinforcement_spend.ts
